refactor: Membership 도메인 관련 API 파일 리패키징
