+++ backend/src/main/java/com/icehufs/icebreaker/controller/UserController.java
@@ -11,9 +11,16 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.icehufs.icebreaker.dto.request.user.*;
-import com.icehufs.icebreaker.dto.response.user.*;
-import com.icehufs.icebreaker.service.UserService;
+import com.icehufs.icebreaker.domain.membership.dto.request.AuthorityRequestDto;
+import com.icehufs.icebreaker.domain.membership.dto.request.PatchUserPassRequestDto;
+import com.icehufs.icebreaker.domain.membership.dto.request.PatchUserRequestDto;
+import com.icehufs.icebreaker.domain.membership.dto.response.Authority1ExistResponseDto;
+import com.icehufs.icebreaker.domain.membership.dto.response.AuthorityResponseDto;
+import com.icehufs.icebreaker.domain.membership.dto.response.DeleteUserResponseDto;
+import com.icehufs.icebreaker.domain.membership.dto.response.GetSignInUserResponseDto;
+import com.icehufs.icebreaker.domain.membership.dto.response.PatchUserPassResponseDto;
+import com.icehufs.icebreaker.domain.membership.dto.response.PatchUserResponseDto;
+import com.icehufs.icebreaker.domain.membership.service.UserService;
 
 import lombok.RequiredArgsConstructor;
